Split library/storage provider.
