+++ mopidy_local/actor.py
@@ -16,7 +16,6 @@ logger = logging.getLogger(__name__)
 
 class LocalBackend(pykka.ThreadingActor, backend.Backend):
     uri_schemes = ['local']
-    libraries = []
 
     def __init__(self, config, audio):
         super(LocalBackend, self).__init__()
@@ -25,15 +24,5 @@ class LocalBackend(pykka.ThreadingActor, backend.Backend):
 
         storage.check_dirs_and_files(config)
 
-        libraries = {l.name: l for l in self.libraries}
-        library_name = config['local']['library']
-
-        if library_name in libraries:
-            library = libraries[library_name](config)
-            logger.debug('Using %s as the local library', library_name)
-        else:
-            library = None
-            logger.warning('Local library %s not found', library_name)
-
         self.playback = LocalPlaybackProvider(audio=audio, backend=self)
-        self.library = LocalLibraryProvider(backend=self, library=library)
+        self.library = LocalLibraryProvider(backend=self, config=config)
