+++ tests/test_tracklist.py
@@ -23,8 +23,14 @@ class LocalTracklistProviderTest(unittest.TestCase):
         },
         'local': {
             'media_dir': path_to_data_dir(''),
-            'playlists_dir': b'',
-            'library': 'json',
+            'base_uri': '/images/',
+            'image_dir': None,
+            'album_art_files': [],
+            'directories': [],
+            'timeout': 10,
+            'use_album_mbid_uri': True,
+            'use_artist_mbid_uri': False,
+            'use_artist_sortname': False
         }
     }
     tracks = [
